app/requirements.txt
altgraph==0.16.1
appdirs==1.4.3
astroid==2.3.0
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
black==19.10b0
bleach==3.1.0
blinker==1.4
bokeh==1.4.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
Flask==1.1.1
Flask-Mail==0.9.1
flask-mongoengine==0.7.5
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.12
future==0.18.2
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.19.1
more-itertools==7.2.0
mplcursors==0.3
mpldatacursor==0.7.1
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==4.4.0
nose==1.3.7
notebook==6.0.2
numpy==1.17.4
openpyxl==3.0.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
passlib==1.7.2
pathspec==0.6.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.2.1
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
pycodestyle==2.5.0
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.4.2
pylint==2.4.1
pymongo==3.10.1
pyparsing==2.4.2
PyQt5==5.13.0
PyQt5-sip==12.7.0
pyqt5-tools==5.13.0.1.5
pyrsistent==0.15.5
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.5.5
regex==2019.11.1
requests==2.22.0
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.0
sip==5.0.0
six==1.12.0
SQLAlchemy==1.3.13
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
yapf==0.28.0
zipp==0.6.0
